+++ foodplanner/lib/components/meal_box.dart
@@ -1,58 +1,97 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_sficon/flutter_sficon.dart';
+import 'package:foodplanner/auth/auth_provider.dart';
 import 'package:foodplanner/components/date_time_picker.dart';
 import 'package:foodplanner/components/meal_box_content.dart';
-import 'package:foodplanner/components/meal_box_empty.dart';
+import 'package:foodplanner/config/colors.dart';
+import 'package:foodplanner/routes/user_roles.dart';
 import 'package:foodplanner/services/meal_notifier.dart';
 import 'package:provider/provider.dart';
 
-class ReusableMealBox extends StatelessWidget {
-  final Size size;
-  final String imageUrl;
-  final String caption;
+class ReusableMealBox extends StatefulWidget {
+  const ReusableMealBox({super.key});
 
-  const ReusableMealBox({Key? key, required this.size})
-      : caption = 'Madpakke Tekst',
-        imageUrl =
-            'https://cdn-icons-png.flaticon.com/512/739/739249.png', // when we fetch we change here so the result is displayed (the picture from minio)
-        super(key: key);
+  @override
+  State<ReusableMealBox> createState() => _ReusableMealBoxState();
+}
+
+class _ReusableMealBoxState extends State<ReusableMealBox> {
+  late MealNotifier mealNotifier;
+
+  Future<ROLES?> _retrieveRole() async {
+    return await AuthProvider().retrieveRole();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    mealNotifier = Provider.of<MealNotifier>(context, listen: false);
+    mealNotifier.fetchMealData();
+  }
 
   @override
   Widget build(BuildContext context) {
-    final mealNotifier = Provider.of<MealNotifier>(context);
-    return Container(
-      width: size.width * 0.9, // 90% of the screen width
-      // if meal empty the grey box is smaller
-      height: mealNotifier.isMealEmpty
-          ? size.height * 0.2
-          : size.height * 0.1 +
-              size.width * 0.6 +
-              190, // "dynamic" height, if mealbox is empty its smaller
-      padding: const EdgeInsets.all(40),
-      decoration: BoxDecoration(
-        color: const Color.fromARGB(
-            255, 243, 243, 243), // image box background color
-        border: Border.all(color: Colors.grey),
-        borderRadius: BorderRadius.circular(40),
-        boxShadow: [
-          // love sabrina carpenter
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            spreadRadius: 5,
-            blurRadius: 7,
-            offset: Offset(0, 3), // changes position of shadow
-          ),
-        ],
+    return Card(
+      color: AppColors.background,
+      surfaceTintColor: AppColors.background,
+      elevation: 2,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(20),
       ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          DateTimePickerWidget(),
-
-          //Added listener for checking if meal is empty or not
-          mealNotifier.isMealEmpty
-              ? MealBoxEmpty(size: size)
-              : MealBoxContent(size: size, caption: caption),
-        ],
+      child: Padding(
+        padding: const EdgeInsets.symmetric(vertical: 20),
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 20),
+              child: FutureBuilder(
+                  future: _retrieveRole(),
+                  builder: (context, snapshot) {
+                    if (snapshot.hasData && snapshot.data == ROLES.parent) {
+                      return InkWell(
+                        onTap: () => mealNotifier.selectDate(context),
+                        overlayColor: WidgetStatePropertyAll(AppColors.primary),
+                        borderRadius: BorderRadius.circular(10),
+                        child: Padding(
+                          padding: const EdgeInsets.only(left: 4),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              DateTimePickerWidget(),
+                              SFIcon(SFIcons.sf_calendar, fontSize: 36),
+                            ],
+                          ),
+                        ),
+                      );
+                    } else {
+                      return Padding(
+                        padding: const EdgeInsets.only(left: 4),
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            DateTimePickerWidget(),
+                            SFIcon(SFIcons.sf_calendar, fontSize: 36),
+                          ],
+                        ),
+                      );
+                    }
+                  }),
+            ),
+            if (mealNotifier.meal == null)
+              Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  // sabrina carpenter tho :flushedEmoj:
+                  const Text(
+                    'ingen madpakke at vise',
+                    style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+                  ),
+                ],
+              )
+            else
+              Mealboxcontent(caption: mealNotifier.meal!.name)
+          ],
+        ),
       ),
     );
   }
