+++ foodplanner/lib/pages/landing_page_parent.dart
@@ -92,7 +92,7 @@ class ParentLandingPageMadpakkeState extends State<ParentLandingPageMadpakke> {
                 children: [
                   Padding(
                     padding: const EdgeInsets.symmetric(horizontal: 20),
-                    child: ReusableMealBox(size: size),
+                    child: ReusableMealBox(),
                   ), // Use the reusable widget
                   SizedBox(height: 20),
                   mealNotifier.meal == null
